+++ src/edu/ucsb/cs56/projects/tutorials/autocomplete/FollowFreq.java
@@ -1,0 +1,10 @@
+package edu.ucsb.cs56.projects.tutorial.autocomplete;
+
+class FollowFreq{
+    public String word;
+    public int freq;
+    
+    public FollowFreq(){word = ""; freq = 0;}
+    public FollowFreq(String word, int freq){ this.word = word; this.freq = freq;}
+    
+};
